# dollmerchant/docs/swagger/swagger.yaml
basePath: /dev/mch/v1
host: 192.168.0.167:9555
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 娃娃机商户平台 Api文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 娃娃机商户平台 Api文档
  version: "1.0"
paths:
...
  /notice/info:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 公告ID
        in: query
        name: Notice_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取公告
      tags:
      - Notice
  /notice/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Token
        in: header
        name: Token
        required: true
        type: string
      - description: 偏移量
        in: query
        name: offset
        required: true
        type: string
      - description: 总条数
        in: query
        name: total_size
        required: true
        type: string
      - description: 一页多少条数据
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取公告列表
      tags:
      - Notice
...
  /test/info:
    put:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Test
        in: formData
        name: test
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 测试专用
      tags:
      - test
...
swagger: "2.0"